+++ vars/hesperides.groovy
@@ -224,6 +224,12 @@ def getDiffProperties(Map args) {
             steps: this.steps).getDiffProperties(args)
 }
 
+def getDiffPropDisplay(Map args) {
+    new Hesperides(apiRootUrl: args.apiRootUrl,
+            auth: args.auth,
+            httpRequester: new JenkinsHTTRequester(this.steps),
+            steps: this.steps).getDiffPropDisplay(args)
+}
 
 /******************************************************************************